added logging features
Added logging features for safety of devices in unmanned, unattended locations. 
Now any access to the serial console or the web interface generate a log message sent via UDP. 
The heartbeats are now sent every 60 seconds, the parameter is hardcoded so if you want to change it, edit esp32_nat_router.c accordingly, and build the firmware.